new backlog folders, files and structure
