+++ backlog/client/add/add.js
@@ -1,0 +1,8 @@
+import { ApiBacklogItems } from '../../collection';
+
+Template.addApiBacklogItem.helpers({
+  apiBacklogItemsCollection () {
+    // Return a reference to ApiBacklogItems collection
+    return ApiBacklogItems;
+  }
+});
