+++ backlog/client/delete/delete.html
@@ -1,0 +1,37 @@
+<template name="deleteBacklogItem">
+  <div
+    class="modal fade"
+    tabindex="-1"
+    role="dialog"
+    aria-labelledby="deleteBacklogItemModalLabel">
+    <div class="modal-dialog" role="document">
+      <div class="modal-content">
+        <div class="modal-header">
+          <button
+            type="button"
+            class="close"
+            data-dismiss="modal"
+            aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          <h2 class="modal-title" id="deleteBacklogItemModalLabel">
+            {{_ "deleteBacklogItem_modalTitle" }}
+          </h2>
+        </div>
+          <div class="modal-body">
+            {{_ "deleteBacklogItem_confirmation_text" }}
+          </div>
+          <div class="modal-footer">
+            <div class="btn-group pull-right">
+              <button id="cancel-delete" class="btn btn-primary" data-dismiss="modal">
+                {{_ "deleteBacklogItem_cancelButton_text" }}
+              </button>
+              <button id="confirm-delete" class="btn btn-danger">
+                {{_ "deleteBacklogItem_deleteButton_text" }}
+              </button>
+            </div>
+          </div>
+      </div>
+    </div>
+  </div>
+</template>
