+++ backlog/client/edit/autoform.js
@@ -1,0 +1,7 @@
+AutoForm.addHooks(['editBacklogItemForm'], {
+  // Success message
+  onSuccess: function () {
+    // Hide Edit Backlog Item modal
+    Modal.hide("editBacklogItem");
+  }
+});
