+++ backlog/client/edit/edit.js
@@ -1,0 +1,8 @@
+import { ApiBacklogItems } from '../../collection';
+
+Template.editBacklogItem.helpers({
+  apiBacklogItemsCollection () {
+    // Return a reference to ApiBacklogItems collection
+    return ApiBacklogItems;
+  }
+});
