+++ backlog/collection/index.js
@@ -1,0 +1,4 @@
+
+const ApiBacklogItems = new Mongo.Collection("apiBacklogItems");
+
+export { ApiBacklogItems };
